requirements.txt
flask==2.0.1
openai==0.27.8
python-dotenv==0.19.0
flask-cors==3.0.10
